postgres_fastapi/backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["MedAziz11"]


[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.67.0"
SQLAlchemy = "^1.4.23"
alembic = "^1.7.1"
uvicorn = "^0.15.0"
psycopg2-binary = "^2.9.1"
pydantic = {extras = ["email", "dotenv"], version = "^1.8.2"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
pytest-cov = "^2.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
